// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  REPORTER: 'REPORTER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ArticleStatus = {
  DRAFT: 'DRAFT',
  PENDING_REVIEW: 'PENDING_REVIEW',
  APPROVED: 'APPROVED',
  PUBLISHED: 'PUBLISHED',
  REJECTED: 'REJECTED',
  ARCHIVED: 'ARCHIVED'
} as const

export type ArticleStatus = (typeof ArticleStatus)[keyof typeof ArticleStatus]


export const CommentStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type CommentStatus = (typeof CommentStatus)[keyof typeof CommentStatus]


export const AdPlacement = {
  HEADER_BANNER: 'HEADER_BANNER',
  SIDEBAR: 'SIDEBAR',
  IN_ARTICLE: 'IN_ARTICLE',
  FOOTER_BANNER: 'FOOTER_BANNER'
} as const

export type AdPlacement = (typeof AdPlacement)[keyof typeof AdPlacement]